refactor(internal): split the variables-ordering pass out of the checker
The ordering pass pushed best_practice_checker.py to 1393 lines, past the
~1000-line mark in AGENTS.md § Module Size. Moved verbatim, no behaviour
change: 939 lines remain in the checker, 383 in best_practice_variables.py.

The warning accumulator, `_emit` and the skill-route helper move to
best_practice_result.py so a check module can emit without importing the
checker back. That is the layer that makes the split acyclic; without it
the new module and the checker would import each other.

The anchor test in test_skill_content_wiring.py parsed
best_practice_checker.py alone. Left as-is it would have kept passing while
silently dropping the one anchor that moved — 25 refs scanned instead of 26,
without the `automation-patterns.md#variables` one. It now globs
best_practice_*.py and asserts the glob is non-empty.

Unit suite unchanged at 8909 passed, 170 skipped, and all 20 mutations from
the previous commit still produce a red test against the moved module.
