+++ src/ha_mcp/tools/best_practice_checker.py
@@ -59,36 +59,14 @@ Anti-patterns sourced from:
 from __future__ import annotations
 
 import re
-from collections.abc import Iterator
 from typing import Any
 
-_SKILL_URI_PREFIX = "skill://home-assistant-best-practices/references"
-_DEFAULT_SKILL_PREFIX = _SKILL_URI_PREFIX
-_SKILL_NAME = "home-assistant-best-practices"
-
-
-class BestPracticeCheckResult(list[str]):
-    """Warning list with an attached set of referenced skill files.
-
-    Behaves as a plain ``list[str]`` for all call sites — ``len()``,
-    indexing, iteration, equality with ``[...]`` all work unchanged.
-    ``referenced_files`` is added on top so the write tools can resolve
-    and embed the relevant skill bodies into responses.
-
-    Use :func:`_emit` to append warnings; direct ``append``/``extend``
-    skip the ``referenced_files`` mirror. Slicing, ``list()`` coercion,
-    and ``copy.copy``/``copy.deepcopy`` all return plain ``list[str]``
-    without the attribute — no current call site does any of these.
-    """
-
-    __slots__ = ("referenced_files",)
-
-    referenced_files: set[str]
-
-    def __init__(self, items: list[str] | None = None) -> None:
-        super().__init__(items or [])
-        self.referenced_files = set()
-
+from .best_practice_result import (
+    _DEFAULT_SKILL_PREFIX,
+    BestPracticeCheckResult,
+    _emit,
+)
+from .best_practice_variables import _check_variables_order
 
 # ---------------------------------------------------------------------------
 # Regex patterns for template anti-patterns
@@ -158,90 +136,6 @@ _RE_MOTION = re.compile(r"binary_sensor\.\w*motion", re.IGNORECASE)
 _RE_ANY_TEMPLATE = re.compile(r"\{\{|\{%")
 # `this.X` self-reference (e.g. `{{ this.entity_id }}`)
 _RE_THIS_REFERENCE = re.compile(r"\bthis\s*\.\s*\w+")
-
-# --- Variables-block ordering scan (see _check_variables_order) -------------
-# Opening delimiter of a Jinja span. Only text inside `{{ }}` / `{% %}` can
-# read a variable, so a plain string value never counts as a reference. The
-# closing delimiter is found by walking forward (see _span_end) instead of by a
-# lazy `.*?`: only a walk that steps over string literals knows that the `}}`
-# in `{{ "}}" ~ later }}` is text and does not end the span.
-_RE_SPAN_OPEN = re.compile(r"\{\{|\{%|\{#")
-_SPAN_CLOSERS = {"{{": "}}", "{%": "%}", "{#": "#}"}
-# Quoted literal inside a span — a sibling's name appearing in one (including
-# as a dict subscript, `x['meldung']`) is text, not a read. Jinja takes Python's
-# string rules, so a backslash escape has to be consumed as one unit or the
-# literal ends early and its tail is scanned as code (`{{ 'don\'t drop x' }}`).
-# ``re.DOTALL`` so an escaped newline inside a literal is consumed with it,
-# matching Jinja's own ``string_re``; without it the literal goes unrecognised
-# and its contents are scanned as code.
-_RE_STRING_LITERAL = re.compile(r"'(?:[^'\\]|\\.)*'|\"(?:[^\"\\]|\\.)*\"", re.DOTALL)
-# Statement tag name, past the optional whitespace-control marker. Jinja
-# accepts `-` and `+` there, so both have to be stepped over.
-_RE_STMT_TAG = re.compile(r"[-+]?\s*([^\W\d]\w*)")
-# `{% set a, b = ... %}` / `{% with a = ... %}` — targets left of the `=`, the
-# expression right of it. Bindings are applied by position, not collected up
-# front: the right-hand side is read *before* the binding takes effect (see
-# _referenced_names). `=(?!=)` so a comparison in the body is not read as an
-# assignment.
-_RE_ASSIGN = re.compile(r"[-+]?\s*(?:set|with)\s+([^=]*?)\s*=(?!=)(.*)", re.DOTALL)
-# `{% set x %}...{% endset %}` — the block form binds without an `=`.
-_RE_BLOCK_SET = re.compile(r"[-+]?\s*set\s+([^\W\d]\w*)\s*[-+]?\s*$")
-# `{% for a, b in expr if cond %}` — the targets bind for the loop body, the
-# iterable is evaluated in the enclosing scope, and the filter clause sees the
-# targets already bound.
-_RE_FOR = re.compile(r"[-+]?\s*for\s+(.*?)\s+in\s+(.*?)(?:\s+if\s+(.*?))?\s*[-+]?$", re.DOTALL)
-# `{% macro name(a, b=1) %}` — the name binds in the enclosing scope, the
-# parameters inside the macro body.
-_RE_MACRO = re.compile(r"[-+]?\s*macro\s+([^\W\d]\w*)\s*\((.*?)\)", re.DOTALL)
-# A bare identifier, used to tell a binding target (`x`) from an assignment
-# into an existing object (`ns.x`, `d[k]`), which reads rather than binds.
-_RE_PLAIN_NAME = re.compile(r"[^\W\d]\w*")
-# `| name` is a filter, never a variable read.
-_RE_FILTER_NAME = re.compile(r"\|\s*[^\W\d]\w*")
-# `is name` / `is not name` — a Jinja test, never a variable read.
-_RE_TEST_NAME = re.compile(r"\bis\s+(?:not\s+)?[^\W\d]\w*")
-# `name=` inside a call is a keyword argument, never a read (`{{ dict(x=1) }}`).
-_RE_KEYWORD_ARG = re.compile(r"[^\W\d]\w*\s*=(?!=)")
-# `.name` is attribute access on whatever sits to its left, not a read of
-# `name`. Stripped as a pass rather than handled by a lookbehind on the
-# identifier, because Jinja also parses the spaced form (`wetter . meldung`)
-# as attribute access and a lookbehind cannot reach across the space.
-_RE_ATTRIBUTE = re.compile(r"\.\s*[^\W\d]\w*")
-# An identifier read. A longer name is handled by the greedy `\w*` alone, which
-# consumes `offene_tueren_extra` whole rather than reporting `offene_tueren`.
-# The lookbehind covers a preceding word char, which only arises after a digit
-# (`e3` in the literal `2.5e3`).
-# `[^\W\d]` rather than `[A-Za-z_]` because Jinja inherits Python's identifier
-# rules, so `über` and `变量` are valid variable names.
-_RE_JINJA_IDENT = re.compile(r"(?<![\w.])[^\W\d]\w*")
-# Jinja's own keywords and literals. A sibling sharing one of these names is
-# never read by the token that spells it. HA's template globals (`states`,
-# `now`, `trigger`, ...) are deliberately NOT listed — that set drifts with HA,
-# and a stale copy would suppress real warnings; it stays a known false
-# positive instead.
-_JINJA_KEYWORDS = frozenset(
-    {
-        "and", "as", "autoescape", "block", "break", "call", "context",
-        "continue", "do", "elif", "else", "endautoescape", "endblock",
-        "endcall", "endfilter", "endfor", "endif", "endmacro", "endraw",
-        "endset", "endtrans", "endwith", "extends", "false", "filter", "for",
-        "from", "if", "ignore", "import", "in", "include", "is", "loop",
-        "macro", "missing", "none", "not", "or", "pluralize", "raw",
-        "recursive", "required", "scoped", "set", "trans", "true", "with",
-        "without", "False", "None", "True",
-    }
-)
-# Tags that open a variable scope, mapped to the tag that closes it. A binding
-# made inside one does not survive it: `{% for %}{% set x = 1 %}{% endfor %}`
-# leaves `x` undefined again, so bindings live on a stack rather than in one
-# flat set.
-_SCOPE_OPENERS = {
-    "for": "endfor",
-    "macro": "endmacro",
-    "with": "endwith",
-    "call": "endcall",
-}
-_SCOPE_CLOSERS = frozenset(_SCOPE_OPENERS.values())
 
 # Target sub-fields scanned for templates. These are the only keys allowed
 # under ``target:`` in HA's modern action schema.
@@ -357,68 +251,6 @@ def check_script_config(
     _check_variables_order(config.get("variables"), warnings, skill_prefix, "variables")
     _dedupe_inplace(warnings)
     return warnings
-
-
-# ---------------------------------------------------------------------------
-# Warning emission + skill-reference helpers
-# ---------------------------------------------------------------------------
-
-
-def _emit(
-    warnings: BestPracticeCheckResult,
-    message: str,
-    skill_prefix: str | None,
-    file_ref: str,
-) -> None:
-    """Append a warning with the 2-route ' See ...' suffix and track the file.
-
-    Args:
-        warnings: Accumulator (also holds ``referenced_files``).
-        message: Human-readable warning body — the inline alternative.
-        skill_prefix: When set, embedded as the ``skill://`` URI route.
-            When ``None``, the URI route is omitted but the other two
-            routes still appear.
-        file_ref: File path relative to the ``references/`` directory of
-            the home-assistant-best-practices skill, optionally with a
-            ``#anchor`` suffix
-            (e.g. ``"automation-patterns.md#native-conditions"``). The
-            anchor is preserved end-to-end: in the ``skill://`` URI for
-            display, and in ``referenced_files`` so the auto-embed path
-            ships only the matching markdown section instead of the
-            whole 10-20 KB reference file.
-    """
-    warnings.append(message + _skill_route_suffix(skill_prefix, file_ref))
-    warnings.referenced_files.add(f"references/{file_ref}")
-
-
-def _skill_route_suffix(skill_prefix: str | None, file_ref: str) -> str:
-    """Build the ' See ...' suffix naming the available skill access routes.
-
-    When ``skill_prefix`` is ``None`` the entire suffix is suppressed —
-    skills are off server-wide, so neither route resolves. Otherwise the
-    suffix names both so the LLM has a working path regardless of which
-    mechanism its client supports:
-
-    1. ``skill://`` URI — for clients that auto-fetch resource URIs.
-       Anchor preserved.
-    2. ``ha_get_skill_guide(skill=..., file=...)`` — explicit tool call,
-       works on every MCP client. Anchor stripped (the tool reads the
-       whole file).
-
-    Auto-embed of the matching section still happens in the next
-    write's response, driven by ``referenced_files``; the
-    ``MandatoryBPS`` opt-out param is not named here by design.
-    """
-    if not skill_prefix:
-        # Skills feature is disabled server-wide; none of the routes work.
-        # Matches the historical no-suffix behaviour.
-        return ""
-    bare_file = f"references/{file_ref.split('#', 1)[0]}"
-    routes = [
-        f"{skill_prefix}/{file_ref}",
-        f"call ha_get_skill_guide(skill={_SKILL_NAME!r}, file={bare_file!r})",
-    ]
-    return " See " + " | ".join(routes)
 
 
 # ---------------------------------------------------------------------------
@@ -846,292 +678,6 @@ def _check_target_dict(
                     skill_prefix,
                     "template-guidelines.md#when-to-avoid-templates",
                 )
-
-
-# ---------------------------------------------------------------------------
-# Variables-block ordering
-# ---------------------------------------------------------------------------
-
-
-def _iter_strings(value: Any) -> Iterator[str]:
-    """Yield every string a variables-block value renders.
-
-    Mirrors ``homeassistant.helpers.template.render_complex``, which recurses
-    into lists and mappings and renders mapping *keys* as well as values.
-    """
-    if isinstance(value, str):
-        yield value
-    elif isinstance(value, list):
-        for item in value:
-            yield from _iter_strings(item)
-    elif isinstance(value, dict):
-        for key, item in value.items():
-            yield from _iter_strings(key)
-            yield from _iter_strings(item)
-
-
-def _span_end(text: str, start: int, closer: str, *, literal_aware: bool) -> int | None:
-    """Return the index of ``closer``, stepping over string literals.
-
-    ``None`` when the span never closes, or when a literal inside it never
-    closes — neither is parseable, and guessing where the span ended is what
-    produces both a hidden read and a phantom one.
-    """
-    index = start
-    while index < len(text):
-        if literal_aware and text[index] in "'\"":
-            literal = _RE_STRING_LITERAL.match(text, index)
-            if literal is None:
-                return None
-            index = literal.end()
-            continue
-        if text.startswith(closer, index):
-            return index
-        index += 1
-    return None
-
-
-def _stmt_tag(body: str) -> str:
-    """Return the tag name opening a ``{% %}`` span, or ``""``."""
-    match = _RE_STMT_TAG.match(body)
-    return match.group(1) if match else ""
-
-
-def _iter_spans(text: str) -> Iterator[tuple[bool, str]]:
-    """Yield ``(is_statement, body)`` per Jinja span, string literals blanked.
-
-    A ``{#...#}`` comment and everything between ``{% raw %}`` and
-    ``{% endraw %}`` yield nothing: neither is code, so a sibling's name in
-    there is not a read.
-    """
-    position = 0
-    in_raw = False
-    while (opening := _RE_SPAN_OPEN.search(text, position)) is not None:
-        opener = opening.group(0)
-        closer = _SPAN_CLOSERS[opener]
-        # A comment body is not code, so a quote inside it binds nothing.
-        end = _span_end(text, opening.end(), closer, literal_aware=opener != "{#")
-        if end is None:
-            return
-        body = _RE_STRING_LITERAL.sub(" ", text[opening.end() : end])
-        position = end + len(closer)
-        if opener == "{#":
-            continue
-        is_statement = opener == "{%"
-        tag = _stmt_tag(body) if is_statement else ""
-        if in_raw:
-            in_raw = tag != "endraw"
-            continue
-        if tag == "raw":
-            in_raw = True
-            continue
-        yield is_statement, body
-
-
-def _read_names(body: str, bound: set[str]) -> set[str]:
-    """Return the identifiers ``body`` reads, minus the ones already bound.
-
-    The strip passes remove the positions where an identifier token is not a
-    variable read at all: a filter name, a test name, a call's keyword
-    argument, and an attribute on the value to its left.
-    """
-    body = _RE_FILTER_NAME.sub(" ", body)
-    body = _RE_TEST_NAME.sub(" ", body)
-    body = _RE_KEYWORD_ARG.sub(" ", body)
-    body = _RE_ATTRIBUTE.sub(" ", body)
-    return {
-        name
-        for name in _RE_JINJA_IDENT.findall(body)
-        if name not in bound and name not in _JINJA_KEYWORDS
-    }
-
-
-def _split_targets(targets: str) -> tuple[set[str], set[str]]:
-    """Split an assignment target list into ``(bound names, names read)``.
-
-    A plain identifier binds. Anything else assigns *into* an existing object
-    (``{% set ns.x = 1 %}``, ``{% set d['k'] = 1 %}``), which reads the base
-    name rather than binding it.
-    """
-    bound: set[str] = set()
-    reads: set[str] = set()
-    for chunk in targets.strip().strip("()[]").split(","):
-        chunk = chunk.strip()
-        if _RE_PLAIN_NAME.fullmatch(chunk):
-            bound.add(chunk)
-        else:
-            reads |= _read_names(chunk, set())
-    return bound, reads
-
-
-def _split_parameters(parameters: str) -> tuple[set[str], set[str]]:
-    """Split a macro parameter list into ``(bound names, names read)``.
-
-    A default value is evaluated where the macro is defined, so it reads from
-    the enclosing scope while the parameter itself binds inside the macro.
-    """
-    bound: set[str] = set()
-    reads: set[str] = set()
-    for chunk in parameters.split(","):
-        name, _, default = chunk.partition("=")
-        name = name.strip()
-        if _RE_PLAIN_NAME.fullmatch(name):
-            bound.add(name)
-        if default:
-            reads |= _read_names(default, set())
-    return bound, reads
-
-
-def _referenced_names(value: Any) -> set[str]:
-    """Return the identifier tokens a variables-block value reads."""
-    names: set[str] = set()
-    for text in _iter_strings(value):
-        # Bindings the template makes itself, innermost scope last. A
-        # `{% set %}` shadows a sibling only from its own position onward, so
-        # the spans are walked in order; collecting targets up front would
-        # swallow a genuine read happening earlier in the string, or in the
-        # binding's own right-hand side. A scope-opening tag pushes a frame
-        # that its closing tag discards again, so a binding made inside a loop
-        # stops shadowing after `{% endfor %}` — which is where the read it
-        # was hiding actually happens.
-        scopes: list[set[str]] = [set()]
-        for is_statement, body in _iter_spans(text):
-            visible = set().union(*scopes)
-            tag = _stmt_tag(body) if is_statement else ""
-
-            if tag in _SCOPE_CLOSERS:
-                if len(scopes) > 1:
-                    scopes.pop()
-                continue
-
-            if tag == "for":
-                match = _RE_FOR.match(body)
-                if match is None:
-                    scopes.append(set())
-                    continue
-                targets, iterable, condition = match.groups()
-                names |= _read_names(iterable, visible)
-                bound, reads = _split_targets(targets)
-                names |= reads - visible
-                scopes.append(bound)
-                if condition:
-                    names |= _read_names(condition, visible | bound)
-                continue
-
-            if tag == "macro":
-                match = _RE_MACRO.match(body)
-                if match is None:
-                    scopes.append(set())
-                    continue
-                scopes[-1].add(match.group(1))
-                bound, reads = _split_parameters(match.group(2))
-                names |= reads - visible
-                scopes.append(bound)
-                continue
-
-            if tag in ("set", "with"):
-                assignment = _RE_ASSIGN.match(body)
-                if assignment is not None:
-                    bound, reads = _split_targets(assignment.group(1))
-                    names |= _read_names(assignment.group(2), visible)
-                    names |= reads - visible
-                else:
-                    block_set = _RE_BLOCK_SET.match(body)
-                    bound = {block_set.group(1)} if block_set else set()
-                if tag == "with":
-                    scopes.append(bound)
-                else:
-                    scopes[-1] |= bound
-                continue
-
-            if tag == "call":
-                names |= _read_names(body, visible)
-                scopes.append(set())
-                continue
-
-            names |= _read_names(body, visible)
-    return names
-
-
-def _check_variables_order(
-    variables: Any,
-    warnings: BestPracticeCheckResult,
-    skill_prefix: str | None,
-    block_key: str,
-) -> None:
-    """Flag a variables key whose template reads a *later* key in the same block.
-
-    HA renders a variables block one entry at a time, feeding each result into
-    the context for the next (``ScriptVariables.async_simple_render`` for an
-    action-level step, ``async_render`` for the top-level and
-    ``trigger_variables`` blocks). A name declared further down is therefore
-    undefined at render time, and what that costs depends on how the template
-    uses it. HA's undefined is ``LoggingUndefined(jinja2.Undefined)``
-    (``helpers/template/__init__.py``), which overrides only ``__str__``,
-    ``__iter__`` and ``__bool__`` to log; every other operation goes through
-    ``_fail_with_undefined_error``:
-
-    * silent — a bare ``{{ later }}``, ``{% if later %}``, ``~`` concatenation,
-      ``==``/``!=``, ``{% for x in later %}`` and ``| length`` yield an empty,
-      falsy, empty-iterating value, so the automation loads and runs on through
-      the wrong branch with only a warning in the log and on the trace
-    * raising — attribute access, item access, arithmetic, an ordering
-      comparison and casts such as ``| int`` raise ``UndefinedError``, log at
-      ERROR and abort the step
-
-    Only *later* siblings count. Names coming from anywhere else — an earlier
-    sibling, an earlier action's ``response_variable``, the trigger context —
-    are legitimate and never flagged.
-
-    Known false positives, deliberately left in rather than paid for with a
-    full Jinja parser:
-
-    * a later sibling whose name also exists in an outer scope, where the
-      reference is legal and resolves outward (the warning text names this one)
-    * a sibling whose name collides with an HA template global (``states``,
-      ``trigger``, ...) or with a filter or test named in a position the strip
-      passes in :func:`_read_names` do not reach. Jinja's own keywords are
-      excluded (:data:`_JINJA_KEYWORDS`); the HA globals are not, because that
-      set drifts with HA and a stale copy would suppress real warnings
-
-    Known gap: when an automation declares both ``trigger_variables`` and
-    ``variables``, HA merges them into a single sequentially-rendered mapping
-    (``components/automation/__init__.py``, ``_create_automation_entities``),
-    so a ``trigger_variables`` key reading a ``variables`` key is a forward
-    read this per-block scan does not see. The reverse direction is safe: the
-    merge puts ``trigger_variables`` first, so a ``variables`` key reading one
-    of them is a backward read. ``trigger_variables`` is additionally rendered
-    on its own when the triggers are attached (``_async_attach_triggers``,
-    ``limited=True``), so a forward read inside that block is real even when
-    a later ``variables`` key overwrites the same name.
-    """
-    if not isinstance(variables, dict) or len(variables) < 2:
-        return
-
-    names = [key for key in variables if isinstance(key, str)]
-    # The last key has no later sibling, so it can never carry a forward read.
-    for index, key in enumerate(names[:-1]):
-        later = set(names[index + 1 :])
-        forward = sorted(later & _referenced_names(variables[key]))
-        if not forward:
-            continue
-        reads = ", ".join(f"`{name}`" for name in forward)
-        _emit(
-            warnings,
-            f"`{block_key}` key `{key}` reads {reads}, declared later in the same "
-            "block — HA renders a variables block one key at a time, so a name "
-            "further down is not defined yet. How that fails depends on the use: "
-            "attribute or item access, arithmetic, an ordering comparison and "
-            "casts like `| int` raise `UndefinedError` and abort the step, while "
-            "a bare `{{ }}`, `{% if %}`, `~`, `==`, iteration and `| length` "
-            "silently yield an empty, falsy value and let the automation run on "
-            "through the wrong branch. Move the keys it reads above it, or split "
-            "them into consecutive `variables:` steps. Not a problem when the "
-            "name is also defined in an outer scope — the reference then "
-            "resolves outward.",
-            skill_prefix,
-            "automation-patterns.md#variables",
-        )
 
 
 # ---------------------------------------------------------------------------
